LU-4856 misc: 64 bit counters for objects tied to memory size
When the number of an object in use or circulation is tied to memory
size of the system, very large memory systems can overflow 32 bit
counters.  This patch converts several of the most obviously exposed
counters - those initialized on the basis of memory size.
This incudes the debug log size, client osc pool/lru, and ldlm pool.
Several additional checks are made to detect overflow on variables
which have not been converted.

+++ libcfs/libcfs/debug.c
@@ -289,7 +289,7 @@ EXPORT_SYMBOL(libcfs_debug_dumplog);
 int libcfs_debug_init(unsigned long bufsize)
 {
 	int    rc = 0;
-	unsigned int max = libcfs_debug_mb;
+	unsigned long max = libcfs_debug_mb;
 
 	init_waitqueue_head(&debug_ctlwq);
